- change rating to accepts double numbers instead of int - change logic of updating showtimes in a way that request contains full list of showtimes and logic replace old with new ones - update entities to contains bi-directional relation with ratings and showtimes - added integration tests for updating showtimes and adding ratings

+++ src/main/kotlin/com/fourthwall/cinemaservice/domain/movie/command/AddMovieRatingCommand.kt
@@ -15,5 +15,5 @@ class AddMovieRatingCommand(
 
     @field:Min(1, message = "Rating must be at least 1")
     @field:Max(5, message = "Rating must be at most 5")
-    val rating: Int
+    val rating: Double
 ) : Command<Movie>
